[UPDATED] Add validation & tests

+++ controllers/taskController.js
@@ -1,7 +1,9 @@
+const moment = require('moment');
 const mongoose = require('mongoose');
 const { ObjectId } = mongoose.Types;
 const Task = require('../models/Task');
 const Project = require('../models/Project');
+const User = require('../models/User');
 
 // Create Task
 exports.createTask = async (req, res) => {
@@ -12,30 +14,50 @@ exports.createTask = async (req, res) => {
     if (!project) return res.status(404).json({ message: 'Project not found' });
 
     // Check if user is part of the project
-    if (!project.members.includes(req.user._id)) {
-      return res.status(403).json({ message: 'Not authorized to add tasks to this project' });
+    if (req.user.role !== 'admin' && !project.members.includes(req.user._id)) {
+      return res.status(403).json({ message: 'Access denied. Not authorized to add tasks to this project' });
     }
+
+    //Check user and user role and task creation permissions 
+    let taskAssignee = req.user._id;
+    if (req.user.role === 'admin' && assignedUser) {
+      const assignedUserExists = await User.findById(assignedUser);
+
+      if (!assignedUserExists) {
+        return res.status(400).json({ message: 'Assigned user does not exist' });
+      }
+
+      if (!project.members.includes(assignedUser)) {
+        project.members.push(assignedUser);
+        await project.save();
+      }
+
+      //Assign Task to the user
+      taskAssignee = assignedUser;
+    }
+
 
     //Check if task already exists with same criteria
     const duplicateTask = await Task.findOne({
       title: title,
       description: description,
       project: projectId,
-      assignedUser: assignedUser,
     });
 
     if (duplicateTask) {
       return res.status(400).json({ message: 'Task already exists' });
     }
 
+    const parsedDueDate = moment(dueDate, "DD/MM/YYYY").toDate();
+
     const task = new Task({
       title,
       description,
       status,
       priority,
-      dueDate,
+      dueDate: parsedDueDate,
       project: projectId,
-      assignedUser,
+      assignedUser: taskAssignee,
     });
 
     await task.save();
@@ -48,7 +70,7 @@ exports.createTask = async (req, res) => {
 
 // Get Tasks - Optional Params (by Project, by User, by Status)
 exports.getTasks = async (req, res) => {
-  const { project, user, status, startDate, endDate, overdue } = req.query;
+  const { project, user, status, startDate, endDate, overdue, priority } = req.query;
 
   try {
 
@@ -57,17 +79,9 @@ exports.getTasks = async (req, res) => {
 
     if (project) {
       if (ObjectId.isValid(project) && typeof project === 'string') {
-        query.project = new ObjectId(project);
+        query.project = project;
       } else {
         return res.status(400).json({ message: 'Invalid Project Id' });
-      }
-    }
-
-    if (user) {
-      if (ObjectId.isValid(user) && typeof user === 'string') {
-        query.assignedUser = new ObjectId(user);
-      } else {
-        return res.status(400).json({ message: 'Invalid User Id' });
       }
     }
 
@@ -75,15 +89,33 @@ exports.getTasks = async (req, res) => {
       query.status = status;
     }
 
+    if (priority) {
+      query.priority = priority;
+    }
+
     if (startDate || endDate) {
+      const parsedStartDate = moment(startDate, "DD/MM/YYYY").toDate();
+      const parsedEndDate = moment(endDate, "DD/MM/YYYY").toDate();
       query.createdAt = {};
-      if (startDate) query.createdAt.$gte = new Date(startDate);
-      if (endDate) query.createdAt.$lte = new Date(endDate);
+      if (startDate) query.createdAt.$gte = new Date(parsedStartDate);
+      if (endDate) query.createdAt.$lte = new Date(parsedEndDate);
     }
 
     if (overdue === 'true') {
       query.dueDate = { $lt: new Date() };
       query.status = { $in: ['In Progress', 'To Do'] }
+    }
+
+    if (req.user.role === 'user') {
+      query.assignedUser = req.user._id;
+    } else if (req.user.role === 'admin') {
+      if (user) {
+        if (ObjectId.isValid(user) && typeof user === 'string') {
+          query.assignedUser = new ObjectId(user);
+        } else {
+          return res.status(400).json({ message: 'Invalid User Id' });
+        }
+      }
     }
 
     const tasks = await Task.find(query)
@@ -101,30 +133,30 @@ exports.getTasksGroupedByUser = async (req, res) => {
     let pipeline = [];
 
     //Group Users
-      pipeline.push({
-        $group: {
-          _id: '$assignedUser',
-          tasks: { $push: '$$ROOT' },
-          count: { $sum: 1 }
-        },
-      })
+    pipeline.push({
+      $group: {
+        _id: '$assignedUser',
+        tasks: { $push: '$$ROOT' },
+        count: { $sum: 1 }
+      },
+    })
 
     //Lookup & Add User Details
-      pipeline.push({
-        $lookup: {
-          from: 'users',
-          localField: '_id',
-          foreignField: '_id',
-          as: 'userDetails'
-        }
-      }, 
+    pipeline.push({
+      $lookup: {
+        from: 'users',
+        localField: '_id',
+        foreignField: '_id',
+        as: 'userDetails'
+      }
+    },
       {
         $unwind: '$userDetails'
-      }, 
+      },
       {
-        $project : {
-          _id: 0, 
-          name: '$userDetails.name', 
+        $project: {
+          _id: 0,
+          name: '$userDetails.name',
           count: 1,
           tasks: 1
         }
@@ -144,21 +176,28 @@ exports.getTasksGroupedByPriority = async (req, res) => {
   try {
     let pipeline = [];
 
-    //Group Users
+    //Filter tasks for non admin users
+    if (req.user.role === 'user') {
       pipeline.push({
-        $group: {
-          _id: '$priority',
-          tasks: { $push: '$$ROOT' },
-          count: { $sum: 1 }
-        },
-      })
+        $match: { assignedUser: req.user._id }
+      });
+    }
+
+    //Group Users
+    pipeline.push({
+      $group: {
+        _id: '$priority',
+        tasks: { $push: '$$ROOT' },
+        count: { $sum: 1 }
+      },
+    })
 
     //Add Priority Type
-      pipeline.push(
+    pipeline.push(
       {
-        $project : {
-          _id: 0, 
-          priority: '$_id', 
+        $project: {
+          _id: 0,
+          priority: '$_id',
           count: 1,
           tasks: 1
         }
@@ -172,24 +211,27 @@ exports.getTasksGroupedByPriority = async (req, res) => {
   }
 };
 
-
 // Update Task
 exports.updateTask = async (req, res) => {
-  const { taskId } = req.params;
+  const { id } = req.params;
   const { title, description, status, dueDate, assignedUser, priority } = req.body;
 
   try {
-    let task = await Task.findById(taskId);
+    let task = await Task.findById(id);
     if (!task) return res.status(404).json({ message: 'Task not found' });
 
-    // Optional: Check if user is authorized to update the task
-    // e.g., task.owner === req.user._id or similar
+    // Check if user is authorized to update the task
+    if (task.assignedUser.toString() !== req.user._id.toString() && req.user.role !== 'admin') {
+      return res.status(403).json({ message: 'Access denied. Not authorized to edit this task' });
+    }
+
+    const parsedDueDate = moment(dueDate, "DD/MM/YYYY").toDate();
 
     if (title) task.title = title;
     if (description) task.description = description;
     if (status) task.status = status;
     if (priority) task.priority = priority;
-    if (dueDate) task.dueDate = dueDate;
+    if (dueDate) task.dueDate = parsedDueDate;
     if (assignedUser) task.assignedUser = assignedUser;
 
     await task.save();
@@ -202,11 +244,15 @@ exports.updateTask = async (req, res) => {
 
 // Delete Task
 exports.deleteTask = async (req, res) => {
-  const { taskId } = req.params;
+  const { id } = req.params;
 
   try {
-    const task = await Task.findById(taskId);
+    const task = await Task.findById(id);
     if (!task) return res.status(404).json({ message: 'Task not found' });
+
+    if (task.assignedUser.toString() !== req.user._id.toString() && req.user.role !== 'admin') {
+      return res.status(403).json({ message: 'Access denied. Not authorized to delete this task' });
+    }
 
     await task.deleteOne();
     res.json({ message: 'Task removed' });
